GC 모드 추가
- Auto: 별도 스레드에서 매 주기마다 알아서 GC 호출
- SpecifiedThreadHint: 주기가 결정되면 애플리케이션에서 Hint()를 호출한 시점에 GC 호출
- Manual: 애플리케이션이 직접 GC를 호출
- FullPurge일 경우 오브젝트 모두 제거할 때까지 대기

+++ Engine/Source/Runtime/Core/Private/Reflection/Property.cpp
@@ -10,7 +10,7 @@ void Property::SetObject(SObject* InThis, SObject* AssignValue) const
 
 SObject* Property::GetObject(SObject* InThis)
 {
-	uint8* Ptr = reinterpret_cast<uint8*>(GetValue<intptr_t>(InThis));
+	uint8* Ptr = reinterpret_cast<uint8*>(GetValue<uint64>(InThis));
 
 	if (!PtrOff.has_value() && Ptr)
 	{
